Разнёс main.js
