+++ js/arrays.js
@@ -1,5 +1,7 @@
 const nicknames = ['Xen','Antik','Chel','Zakat','Печенька','Чешир','Alukard','СуперВасяxXx','Куница','Че'];
 const descriptions = ['Описание 1', 'Описание 2','Описание 3','Описание 4','Описание 5'];
 const messages = ['Всё отлично!','В целом всё неплохо. Но не всё.','Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+const ids = [];
+const photos = [];
 
-export {nicknames, descriptions, messages};
+export {nicknames, descriptions, messages, ids, photos};
